A new class for creating and parsing tab-separated variable-header text files has been added.
The class is used in the CreateReadmeFileStep step in the px-publication part of the pipeline. This step is now half completed. The README file is created just with the header. We need to be able to obtain the accession numbers from the submission summary file before continue (or find an alternative solution).